Converted AccountController to Use Models
Converted the AccountController from using Entities to using models.
Fixed a few bugs I found with the BO layer of setting up accounts and
have the daily intake controller commented out for now. Updated the
models to copy the IDs as well.. They are needed for the update/create
procedures. If it isn't there, it will try to create it which causes
duplicate key problems n stuffs.