Renaming Packages and Entity Stuff
Changed the location of most of the objects. Moved the entities to
model.entity and removed all the Entity suffuxes. MOved the existing
Front End Models to model.view and kept the Model suffux. Moved the
Contracts to just model. Then moved the controller and navigation
packages to com.mathtabolism.view for a better grouping.