Fixed Daily Intake Problems
Fixed the daily intake problems for the page. I hand't finsihed the
whole saving/creating of all the stupid nested lists of elements in the
daily intake crap. Really need the EntityModelConverter to
auto-magically convert nested entities or models as well. Shouldn't be
too bad. Just check if the parameter type is com.mathtabolism then do a
convert on dat. SON