Formatting, UML, and Fixed Account Indicaotr
Changed some of the formatting of code. Updated the toString methods to
have the entity id. Added a UML diagram to help me understand the
DailyIntake flow. Moved the useBirthday indicator from the Account to
the AccountSettings and renamed it to useAge. I did this because
somebody might want to change this later on. However. I just realized
this can cause calculation errors.. Sigh.