Step Form Fixes
Added some fixes for ajax to work. It isn't really the most ideal way of
doing it so far.. Oh well. But we have the first 2 steps. The height
stuff doesn't update correclty when switching tabs. It saves in the
model correctly, but doesnt display the right value. Also multiple items
are selected. Weird.