Finished the Account Initialization
Fixed and finished the account initialization for now. Just need to
change the height stuff to be input boxes for metric since the
centimeters should go to 100 not 10.. woops. Also fixed some weight
stuff. Had forgotten to give a Temporal type for the weigh in date, so
was unable to find by date. Stupid mistake.