Passed All Entity To Model Tests
Ended up changing the idea from EntitiesToModel to just entityToModel(E
entity, M model) where a new model would be generated if the given model
was null. This is just because there is a possible heap problem when
passing in a varags generic to a method. I could have just ignored the
message. Oh well.